# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "BenchConsole.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/devices": {
      "get": {
        "tags": [
          "Benches"
        ],
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "model",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BenchDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BenchDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BenchDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Benches"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBenchRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBenchRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBenchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BenchDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/devices/{code}": {
      "get": {
        "tags": [
          "Benches"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BenchDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchDto"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Benches"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBenchRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBenchRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBenchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BenchDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Benches"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/devices/{code}/start": {
      "post": {
        "tags": [
          "Benches"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartTestRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/StartTestRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/StartTestRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAcceptedDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAcceptedDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAcceptedDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/devices/{code}/stop": {
      "post": {
        "tags": [
          "Benches"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "by",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAcceptedDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAcceptedDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAcceptedDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/devices/{code}/reset": {
      "post": {
        "tags": [
          "Benches"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "by",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAcceptedDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAcceptedDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAcceptedDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/devices/{code}/deploy": {
      "post": {
        "tags": [
          "Benches"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrienKhaiGoiRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TrienKhaiGoiRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TrienKhaiGoiRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAcceptedDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAcceptedDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAcceptedDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/devices/{code}/telemetry": {
      "get": {
        "tags": [
          "Benches"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "channel",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "minutes",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 5
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TelemetrySeriesDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TelemetrySeriesDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TelemetrySeriesDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BenchDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "code": {
            "type": "string",
            "nullable": true
          },
          "model": {
            "type": "string",
            "nullable": true
          },
          "workshop": {
            "type": "string",
            "nullable": true
          },
          "rack": {
            "type": "string",
            "nullable": true
          },
          "firmware": {
            "type": "string",
            "nullable": true
          },
          "tenMay": {
            "type": "string",
            "nullable": true
          },
          "state": {
            "type": "string",
            "nullable": true
          },
          "note": {
            "type": "string",
            "nullable": true
          },
          "testCase": {
            "type": "string",
            "nullable": true
          },
          "plan": {
            "type": "string",
            "nullable": true
          },
          "step": {
            "type": "string",
            "nullable": true
          },
          "primaryChannel": {
            "type": "string",
            "nullable": true
          },
          "primaryValue": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "primaryUnit": {
            "type": "string",
            "nullable": true
          },
          "lastSeenAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "staleSeconds": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}